Add options to set merge output format and sorting format
yt-dlp uses mp4 format by default to merge sources with separate video and
audio streams. mp4 format does not handle VP9 codec, so adding the ability
to set the merge output format, we can set it to use for ex. mkv, which
handles VP9 well.

With the possibility of setting the sorting format we can tell yt-dlp which
criteria to consider as the best format.