Add the NoInfoDownload option
Set it to true if you don't want to use the result.Info structure.

If it is set to true, then the New() call won't call youtube-dl to request info
about the source in JSON format, and result.Info structure will not be
populated. The given URL will be downloaded in a single pass in the Download()
call.